requirements.txt
pytest>=8.2.0
playwright>=1.44.0
pytest-playwright>=0.5.0
pytest-xdist>=3.6.0
allure-pytest>=2.13.5
python-dotenv>=1.0.1
pydantic>=2.6.0
ruff>=0.5.0
black>=24.4.0
